+++ lib/rubocop/cop/style/symbol_proc.rb
@@ -74,7 +74,11 @@ module RuboCop
         end
 
         def autocorrect_with_args(corrector, node, args, method_name)
-          corrector.insert_after(args.last.source_range, ", &:#{method_name}")
+          arg_range = args.last.source_range
+          arg_range = range_with_surrounding_comma(arg_range, :right)
+          replacement = " &:#{method_name}"
+          replacement = ',' + replacement unless arg_range.source.end_with?(',')
+          corrector.insert_after(arg_range, replacement)
           corrector.remove(block_range_with_space(node))
         end